requirements.txt
pymongo~=4.0.1
python-dotenv~=0.19.2
setuptools~=57.0.0
fastapi~=0.70.0
pytest~=6.2.5
requests~=2.26.0
uvicorn~=0.15.0